requirements.txt
python-dotenv==1.0.0
python-dateutil==2.8.2
Werkzeug==2.3.6
zipp==3.15.0
setuptools==58.1.0
Flask==2.3.2
Jinja2==3.1.2
MarkupSafe==2.1.3
colorama==0.4.6
click==8.1.3
blinker==1.6.2
numpy==1.25.0
pandas==2.0.3
tensorflow==2.12.0